+++ R/bayesfactor_parameters.R
@@ -118,7 +118,6 @@ bayesfactor_parameters <- function(posterior, prior = NULL, direction = "two-sid
 bayesfactor_savagedickey <- function(posterior, prior = NULL, direction = "two-sided", null = 0, verbose = TRUE, hypothesis = NULL, ...) {
   .Deprecated("bayesfactor_parameters")
 
-  dots <- list(...)
   if (!is.null(hypothesis)) {
     null <- hypothesis
     warning("The 'hypothesis' argument is deprecated. Please use 'null' instead.")