+++ R/rope_range.R
@@ -73,6 +73,13 @@ rope_range.lm <- rope_range.brmsfit
 #' @export
 rope_range.merMod <- rope_range.brmsfit
 
+#' @export
+rope_range.sim.merMod <- function(x, ...) {
+  c(-.1, .1)
+}
+
+#' @export
+rope_range.sim <- rope_range.sim.merMod
 
 
 .rope_range <- function(x, information, response) {