+++ R/utils_hdi_ci.R
@@ -178,3 +178,40 @@
   class(dat) <- unique(c(paste0("bayestestR_", fun), paste0("see_", fun), class(dat)))
   dat
 }
+
+
+
+#' @keywords internal
+.compute_interval_simMerMod <- function(x, ci, effects, parameters, verbose, fun) {
+  fixed <- fixed.data <- NULL
+  random <- random.data <- NULL
+
+  if (effects %in% c("fixed", "all")) {
+    fixed.data <- insight::get_parameters(x, effects = "fixed", parameters = parameters)
+    fixed <- .compute_interval_dataframe(fixed.data, ci, verbose, fun)
+    fixed$Group <- "fixed"
+  }
+
+  if (effects %in% c("random", "all")) {
+    random.data <- insight::get_parameters(x, effects = "random", parameters = parameters)
+    random <- .compute_interval_dataframe(random.data, ci, verbose, fun)
+    random$Group <- "random"
+  }
+
+  d <- do.call(rbind, list(fixed, random))
+
+  if (length(unique(d$Group)) == 1) {
+    d <- .remove_column(d, "Group")
+  }
+
+  list(result = d, data = do.call(cbind, .compact_list(list(fixed.data, random.data))))
+}
+
+
+
+#' @keywords internal
+.compute_interval_sim <- function(x, ci, parameters, verbose, fun) {
+  fixed.data <- insight::get_parameters(x, parameters = parameters)
+  d <- .compute_interval_dataframe(fixed.data, ci, verbose, fun)
+  list(result = d, data = fixed.data)
+}