+++ paper/Figure1.R
@@ -1,6 +1,6 @@
 library(bayestestR)
 library(ggplot2)
-library(strengejacke)
+# library(strengejacke)
 
 set.seed(123)
 posterior <- distribution_chisquared(100, 3)
@@ -27,12 +27,12 @@ label_y3 <- ypos3 + .01
 ggplot(dat, aes(x=x, y=y)) +
   geom_ribbon(aes(ymin=0, ymax=y), fill="#FFC107") +
   geom_vline(xintercept=0, linetype="dotted") +
-  geom_segment(x=m, xend=m, y=0, yend=ypos, color="#E91E63", size=1) +
-  geom_point(x=m, y=ypos, color="#E91E63", size=5) +
-  geom_label(x = label_x, y = label_y, label = "Median", color="#E91E63", size = 7) +
-  geom_segment(x=m2, xend=m2, y=0, yend=ypos2, color="#2196F3", size=1) +
-  geom_point(x=m2, y=ypos2, color="#2196F3", size=5) +
-  geom_label(x = label_x2, y = label_y2, label = "MAP", color="#2196F3", size = 7) +
+  geom_segment(x=m2, xend=m2, y=0, yend=ypos2, color="#E91E63", size=1) +
+  geom_point(x=m2, y=ypos2, color="#E91E63", size=5) +
+  geom_label(x = label_x2, y = label_y2, label = "MAP", color="#E91E63", size = 7) +
+  geom_segment(x=m, xend=m, y=0, yend=ypos, color="#2196F3", size=1) +
+  geom_point(x=m, y=ypos, color="#2196F3", size=5) +
+  geom_label(x = label_x, y = label_y, label = "Median", color="#2196F3", size = 7) +
   geom_segment(x=m3, xend=m3, y=0, yend=ypos3, color="#4CAF50", size=1) +
   geom_point(x=m3, y=ypos3, color="#4CAF50", size=5) +
   geom_label(x = label_x3, y = label_y3, label = "Mean", color="#4CAF50", size = 7) +
@@ -41,4 +41,4 @@ ggplot(dat, aes(x=x, y=y)) +
   xlab("\nParameter Value") +
   ylab("Probability Density\n")
 
-ggsave("Figure1.png", width = 13, height = 8, units = "in", dpi = 450)
+ggsave("Figure1b.png", width = 13, height = 8, units = "in", dpi = 450)