+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## Changes to functions
 
-* Added suppoirt for `sim` and `sim.merMod`-objects (from `arm::sim()`) to many functions (like `hdi()`, `ci()`, `eti()`, `rope()`, `p_direction()`, `point_estimate()`, `bayesfactor()`, ...).
+* Added support for `sim` and `sim.merMod`-objects (from `arm::sim()`) to many functions (like `hdi()`, `ci()`, `eti()`, `rope()`, `p_direction()`, `point_estimate()`, `bayesfactor()`, ...).
 
 # bayestestR 0.2.5